+++ src/Network/Network/VirtualRouterPeer/UpdateAzureRMVirtualRouterPeerCommand.cs
@@ -86,7 +86,7 @@ namespace Microsoft.Azure.Commands.Network
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "The virtual router peer resource Id.")]
         [ValidateNotNullOrEmpty]
-        [ResourceIdCompleter("Microsoft.Network/virtualRouters/peerings")]
+        [ResourceIdCompleter("Microsoft.Network/virtualHubs/bgpConnections")]
         public string ResourceId { get; set; }
 
         [Parameter(
@@ -114,11 +114,11 @@ namespace Microsoft.Azure.Commands.Network
             base.Execute();
 
             var present = true;
-            VirtualRouterPeering existingVirtualRouterPeer = null;
+            BgpConnection existingBgpConnection = null;
 
             try
             {
-                existingVirtualRouterPeer = this.NetworkClient.NetworkManagementClient.VirtualRouterPeerings.Get(this.ResourceGroupName, this.VirtualRouterName, this.PeerName);
+                existingBgpConnection = this.NetworkClient.NetworkManagementClient.VirtualHubBgpConnection.Get(this.ResourceGroupName, this.VirtualRouterName, this.PeerName);
             }
             catch (Microsoft.Rest.Azure.CloudException exception)
             {
@@ -141,8 +141,8 @@ namespace Microsoft.Azure.Commands.Network
             if (ParameterSetName.Equals(VirtualRouterPeerParameterSetNames.ByVirtualRouterPeerResourceId, StringComparison.OrdinalIgnoreCase))
             {
                 var parsedResourceId = new ResourceIdentifier(ResourceId);
-                PeerAsn = (uint) existingVirtualRouterPeer.PeerAsn;
-                PeerIp = existingVirtualRouterPeer.PeerIp;
+                PeerAsn = (uint)existingBgpConnection.PeerAsn;
+                PeerIp = existingBgpConnection.PeerIp;
             }
 
             ConfirmAction(
@@ -151,21 +151,27 @@ namespace Microsoft.Azure.Commands.Network
                 () =>
                 {
                     WriteVerbose(String.Format(Properties.Resources.UpdatingLongRunningOperationMessage, this.ResourceGroupName, this.PeerName));
-                    PSVirtualRouterPeer virtualRouterPeer = new PSVirtualRouterPeer
+                    PSVirtualRouterPeer peer = new PSVirtualRouterPeer
                     {
                         Name = this.PeerName,
                         PeerAsn = this.PeerAsn,
                         PeerIp = this.PeerIp
                     };
+                    string ipConfigName = "ipconfig1";
 
-                    var vVirtualRouterPeerModel = NetworkResourceManagerProfile.Mapper.Map<MNM.VirtualRouterPeering>(virtualRouterPeer);
+                    
+                    var bgpConnectionModel = NetworkResourceManagerProfile.Mapper.Map<MNM.BgpConnection>(peer);
 
-                    this.NetworkClient.NetworkManagementClient.VirtualRouterPeerings.CreateOrUpdate(this.ResourceGroupName, this.VirtualRouterName, this.PeerName, vVirtualRouterPeerModel);
-                    var getVirtualRouter = this.NetworkClient.NetworkManagementClient.VirtualRouters.Get(this.ResourceGroupName, this.VirtualRouterName);
-                    var vVirtualRouterModel = NetworkResourceManagerProfile.Mapper.Map<CNM.PSVirtualRouter>(getVirtualRouter);
-                    vVirtualRouterModel.ResourceGroupName = this.ResourceGroupName;
-                    vVirtualRouterModel.Tag = TagsConversionHelper.CreateTagHashtable(getVirtualRouter.Tags);
-                    AddPeeringsToPSVirtualRouter(getVirtualRouter, vVirtualRouterModel, ResourceGroupName, this.VirtualRouterName);
+                    this.NetworkClient.NetworkManagementClient.VirtualHubBgpConnection.CreateOrUpdate(this.ResourceGroupName, this.VirtualRouterName, this.PeerName, bgpConnectionModel);
+                    var virtualHub = this.NetworkClient.NetworkManagementClient.VirtualHubs.Get(this.ResourceGroupName, this.VirtualRouterName);
+                    var virtualHubModel = NetworkResourceManagerProfile.Mapper.Map<CNM.PSVirtualHub>(virtualHub);
+                    virtualHubModel.ResourceGroupName = this.ResourceGroupName;
+                    virtualHubModel.Tag = TagsConversionHelper.CreateTagHashtable(virtualHub.Tags);
+                    AddBgpConnectionsToPSVirtualHub(virtualHub, virtualHubModel, ResourceGroupName, this.VirtualRouterName);
+                    AddIpConfigurtaionToPSVirtualHub(virtualHubModel, this.ResourceGroupName, this.VirtualRouterName, ipConfigName);
+
+                    var vVirtualRouterModel = new PSVirtualRouter(virtualHubModel);
+                    vVirtualRouterModel.Tag = TagsConversionHelper.CreateTagHashtable(virtualHub.Tags);
 
                     WriteObject(vVirtualRouterModel, true);
                 });