+++ src/Synapse/Synapse/Common/HelpMessages.cs
@@ -307,5 +307,13 @@ namespace Microsoft.Azure.Commands.Synapse.Common
         public const string TriggerName = "The trigger name.";
 
         public const string TriggerObject = "The trigger object.";
+
+        public const string DatasetName = "The dataset name.";
+
+        public const string DatasetObject = "The dataset object.";
+
+        public const string DataFlowName = "The data flow name.";
+
+        public const string DataFlowObject = "The data flow object.";
     }
 }