+++ src/Aks/Aks/ChangeLog.md
@@ -18,7 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Add parameter `GenerateSshKey` for `New-AzAksCluster`.
+* Added parameter `GenerateSshKey` for `New-AzAksCluster`.
+* Updated api version to 2020-06-01.
 
 ## Version 1.2.0
 * Removed `ClientIdAndSecret` to `ServicePrincipalIdAndSecret` and set `ClientIdAndSecret` as an alias [#12381].