+++ src/Aks/Aks/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added client side parameter validation logic for `New-AzAksCluster`, `Set-AzAksCluster` and `New-AzAksNodePool`.
 * Added parameter `GenerateSshKey` for `New-AzAksCluster`.
 * Updated api version to 2020-06-01.